[pigeon] Encode custom enum as long in Kotlin generator

+++ packages/pigeon/platform_tests/test_plugin/android/src/main/kotlin/com/example/test_plugin/CoreTests.gen.kt
@@ -704,11 +704,11 @@ private open class CoreTestsPigeonCodec : StandardMessageCodec() {
     when (value) {
       is AnEnum -> {
         stream.write(129)
-        writeValue(stream, value.raw)
+        writeValue(stream, value.raw.toLong())
       }
       is AnotherEnum -> {
         stream.write(130)
-        writeValue(stream, value.raw)
+        writeValue(stream, value.raw.toLong())
       }
       is UnusedClass -> {
         stream.write(131)
